requirements.txt
amqp
asgiref
backports.zoneinfo
billiard
celery
certifi
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
Django
django-appconf
django-celery-email
djangorestframework
flower
humanize
idna
kombu
Pillow
prometheus-client
prompt-toolkit
psycopg2
python-dateutil
pytz
requests
six
sqlparse
tornado
typing_extensions
tzdata
urllib3
vine
watchdog
wcwidth
djoser
